requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
gunicorn
djangorestframework
dnspython
email-validator
aiosmtplib
pymemcache
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
dnspython==2.8.0
email-validator==2.3.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
sqlparse==0.5.3
celery[redis]
